requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssbeautifier==1.15.1
daphne==4.1.2
defusedxml==0.7.1
distro==1.9.0
Django==5.1.4
django-bootstrap-icons==0.9.0
django-bootstrap5==24.3
djlint==1.36.3
EditorConfig==0.17.0
flatbuffers==24.12.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
jiter==0.8.2
jsbeautifier==1.15.1
json5==0.10.0
kagglehub==0.3.6
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
milvus-lite==2.4.11
ml-dtypes==0.4.1
model-bakery==1.20.0
msgpack==1.1.0
mypy-extensions==1.0.0
namex==0.0.8
numpy==2.0.2
openai==1.59.7
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.3
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-text==2.18.1
termcolor==2.5.0
tf_keras==2.18.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zope.interface==7.2
